+++ src/targets.py
@@ -87,8 +87,8 @@ class Target:
         else:
             return self.url.host
 
-    def __hash__(self):
-        return hash(self.url)
+    def create_stats(self, method: str) -> "TargetStats":
+        return TargetStats(self, method)
 
 
 class TargetsLoader:
@@ -144,3 +144,34 @@ class TargetsLoader:
 
         return targets
 
+
+class TargetStats:
+
+    def __init__(self, target: Target, method: str):
+        self._target = target
+        self._method = method
+        self._sig = target.options_repr
+        self._requests: int = 0
+        self._bytes: int = 0
+        self._conns: int = 0
+        self._reset_at = time.perf_counter()
+
+    @property
+    def target(self) -> Tuple[Target, str, str]:
+        return (self._target, self._method, self._sig)
+
+    def track(self, rs: int, bs: int) -> None:
+        self._requests += rs
+        self._bytes += bs
+
+    def track_open_connection(self) -> None:
+        self._conns += 1
+
+    def track_close_connection(self) -> None:
+        self._conns -= 1
+
+    def reset(self) -> Tuple[int, int, int]:
+        sent_requests, sent_bytes, prev_reset_at = self._requests, self._bytes, self._reset_at
+        self._requests, self._bytes, self._reset_at = 0, 0, time.perf_counter()
+        interval = self._reset_at - prev_reset_at
+        return int(sent_requests / interval), int(8 * sent_bytes / interval), self._conns
